+++ src/qq_lib/batch/pbs/job.py
@@ -10,7 +10,9 @@ from typing import Self
 import yaml
 
 from qq_lib.batch.interface import BatchJobInterface
-from qq_lib.batch.pbs.common import parse_pbs_dump_to_dictionary
+from qq_lib.batch.pbs.common import (
+    parse_pbs_dump_to_dictionary,
+)
 from qq_lib.core.common import hhmmss_to_duration, load_yaml_dumper
 from qq_lib.core.config import CFG
 from qq_lib.core.error import QQError
@@ -70,6 +72,11 @@ class PBSJob(BatchJobInterface):
     def getState(self) -> BatchState:
         if not (state := self._info.get("job_state")):
             return BatchState.UNKNOWN
+
+        # X is used by PBS to indicate finished tasks in unfinished array jobs,
+        # but qq uses X to indicate failure
+        if state == "X":
+            state = "F"
 
         # if the job is finished and the return code is not zero, return FAILED
         if state == "F":